Introduce proxymate, TCP to TLS/WebSockets proxy
proxymate is a simple plain-TCP to TLS and/or WebSockets
(or vice versa) proxy written in async Rust.

It leverages the 'rustls' and 'tungstenite' crates and their async
counterparts 'async_tls' and 'async_tungstenite' respectively.

For now the actual data transfer between a client and the
target is implemented by simply copying stream of bytes from
AsyncRead to AsyncWrite.

While this choice allows the proxy to handle all three kinds of
sockets (TCP, TLS, WS) in a uniform way, there might be some
drawbacks speaking of WebSockets stream, namely:
 - WS is message oriented, therefore any information regarding
   message type, error, control, etc. are lost, only data bytes
   are transfered
 - There may be performance penalty due to double buffering
   in layer that handles conversion between messages stream to
   Async{Read,Write}
 - The current version of ws_stream_tungstenite crate provides
   support only for binary messages and any received text message
   is treated as an error that results in disconnect.